+++ srcs/ft_toplace.c
@@ -6,13 +6,13 @@
 /*   By: ybenoit <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/21 14:30:22 by ybenoit           #+#    #+#             */
-/*   Updated: 2016/11/21 14:39:43 by ybenoit          ###   ########.fr       */
+/*   Updated: 2016/11/23 12:02:30 by arive-de         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fillit.h"
 
-t_map	*ft_toplace(t_map *mymap, char **pattern, int x, int y)
+t_map	*ft_toplace(t_map *mymap, char **pattern, size_t x, size_t y)
 {
 	if (ft_ifplace(mymap, pattern, x, y))
 	{
